# servers/core/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a core sever of PROMPT.",
        "title": "PROMPT Core API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/",
    "paths": {
...
        "/course_phases/{uuid}/participations/students": {
            "get": {
                "description": "Get all students of a given course phase",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "course_phase_participation"
                ],
                "summary": "Get students of a course phase",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course Phase UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/studentDTO.Student"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/keycloak/{courseID}/group/editor/students": {
            "put": {
                "description": "Add students to the editor group for a course",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "keycloak"
                ],
                "summary": "Add students to the editor group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course UUID",
                        "name": "courseID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Students to add",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/keycloakRealmDTO.AddStudentsToGroup"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/keycloakRealmDTO.AddStudentsToGroupResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/keycloak/{courseID}/group/{groupName}/students": {
            "get": {
                "description": "Get all students in a specific group for a course",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "keycloak"
                ],
                "summary": "Get students in a group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course UUID",
                        "name": "courseID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Group name",
                        "name": "groupName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/keycloakRealmDTO.GroupMembers"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Add students to a custom group for a course",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "keycloak"
                ],
                "summary": "Add students to a custom group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course UUID",
                        "name": "courseID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Group name",
                        "name": "groupName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Students to add",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/keycloakRealmDTO.AddStudentsToGroup"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/keycloakRealmDTO.AddStudentsToGroupResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/students/": {
            "get": {
                "description": "Get a list of all students",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "students"
                ],
                "summary": "Get all students",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/studentDTO.Student"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new student",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "students"
                ],
                "summary": "Create a student",
                "parameters": [
                    {
                        "description": "Student to create",
                        "name": "student",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/studentDTO.CreateStudent"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/studentDTO.Student"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/students/search/{searchString}": {
            "get": {
                "description": "Search students by a search string",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "students"
                ],
                "summary": "Search students",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search string",
                        "name": "searchString",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/studentDTO.Student"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/students/with-courses": {
            "get": {
                "description": "Get a list of all students with the property 'courses' a list of courses that the student is taking part of or was",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "students"
                ],
                "summary": "Get all students with courses",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/studentDTO.StudentWithCourseParticipationsDTO"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/students/{uuid}": {
            "get": {
                "description": "Get a student by UUID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "students"
                ],
                "summary": "Get student by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Student UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/studentDTO.Student"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing student by UUID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "students"
                ],
                "summary": "Update a student",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Student UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Student to update",
                        "name": "student",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/studentDTO.CreateStudent"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/studentDTO.Student"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/students/{uuid}/enrollments": {
            "get": {
                "description": "Get all of a students enrollments, provide student UUID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "students"
                ],
                "summary": "Get student enrollments by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Student UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/studentDTO.StudentEnrollmentsDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "db.Gender": {
            "type": "string",
            "enum": [
                "male",
                "female",
                "diverse",
                "prefer_not_to_say"
            ],
            "x-enum-varnames": [
                "GenderMale",
                "GenderFemale",
                "GenderDiverse",
                "GenderPreferNotToSay"
            ]
        },
...
        "db.StudyDegree": {
            "type": "string",
            "enum": [
                "bachelor",
                "master"
            ],
            "x-enum-varnames": [
                "StudyDegreeBachelor",
                "StudyDegreeMaster"
            ]
        },
        "keycloakRealmDTO.AddStudentsToGroup": {
            "type": "object",
            "properties": {
                "studentsToAdd": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "keycloakRealmDTO.AddStudentsToGroupResponse": {
            "type": "object",
            "properties": {
                "failedToAddStudentIDs": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "succeededToAddStudentIDs": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "keycloakRealmDTO.GroupMembers": {
            "type": "object",
            "properties": {
                "nonStudents": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/keycloakRealmDTO.KeycloakUser"
                    }
                },
                "students": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/studentDTO.Student"
                    }
                }
            }
        },
        "keycloakRealmDTO.KeycloakUser": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "studentDTO.CourseEnrollmentDTO": {
            "type": "object",
            "properties": {
                "courseId": {
                    "type": "string"
                },
                "courseParticipationId": {
                    "type": "string"
                },
                "coursePhases": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/studentDTO.CoursePhaseEnrollmentDTO"
                    }
                },
                "courseType": {
                    "type": "string"
                },
                "ects": {
                    "type": "integer"
                },
                "endDate": {
                    "type": "string"
                },
                "longDescription": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "semesterTag": {
                    "type": "string"
                },
                "startDate": {
                    "type": "string"
                },
                "studentReadableData": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "studentDTO.CoursePhaseEnrollmentDTO": {
            "type": "object",
            "properties": {
                "coursePhaseId": {
                    "type": "string"
                },
                "coursePhaseType": {
                    "$ref": "#/definitions/studentDTO.CoursePhaseTypeDTO"
                },
                "isInitialPhase": {
                    "type": "boolean"
                },
                "lastModified": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "passStatus": {
                    "type": "string"
                }
            }
        },
        "studentDTO.CoursePhaseTypeDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "studentDTO.CreateStudent": {
            "type": "object",
            "properties": {
                "currentSemester": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "gender": {
                    "$ref": "#/definitions/db.Gender"
                },
                "hasUniversityAccount": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "matriculationNumber": {
                    "type": "string"
                },
                "nationality": {
                    "type": "string"
                },
                "studyDegree": {
                    "$ref": "#/definitions/db.StudyDegree"
                },
                "studyProgram": {
                    "type": "string"
                },
                "universityLogin": {
                    "type": "string"
                }
            }
        },
        "studentDTO.Student": {
            "type": "object",
            "properties": {
                "currentSemester": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "gender": {
                    "$ref": "#/definitions/db.Gender"
                },
                "hasUniversityAccount": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "matriculationNumber": {
                    "type": "string"
                },
                "nationality": {
                    "type": "string"
                },
                "studyDegree": {
                    "$ref": "#/definitions/db.StudyDegree"
                },
                "studyProgram": {
                    "type": "string"
                },
                "universityLogin": {
                    "type": "string"
                }
            }
        },
        "studentDTO.StudentCourseParticipationDTO": {
            "type": "object",
            "properties": {
                "courseId": {
                    "type": "string"
                },
                "courseName": {
                    "type": "string"
                },
                "studentReadableData": {
                    "type": "object",
                    "additionalProperties": {}
                }
            }
        },
        "studentDTO.StudentEnrollmentsDTO": {
            "type": "object",
            "properties": {
                "courses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/studentDTO.CourseEnrollmentDTO"
                    }
                }
            }
        },
        "studentDTO.StudentWithCourseParticipationsDTO": {
            "type": "object",
            "properties": {
                "courses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/studentDTO.StudentCourseParticipationDTO"
                    }
                },
                "currentSemester": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "hasUniversityAccount": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "studyProgram": {
                    "type": "string"
                }
            }
        },
        "utils.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Something went wrong"
                }
            }
        }
    },
...
}